Rename `AdBlockingRecoveryCTA` to `AdBlockingRecoverySetupCTANotice`.
